Corrige Grover() medindo o resultado certo e descartando (item 12)
Grover() montava "result" a partir de dgm.measure() -- exatamente como
docs/05-algoritmo-grover.md ja descrevia -- mas so retornava "float
elapsed" (tempo decorrido). grover.cpp so imprimia esse tempo. Das
duas implementacoes de algoritmo do projeto, so o Shor reportava se
tinha encontrado a resposta certa; rodar grover.out nunca dizia nada
alem de "nao crashou", mesmo o programa ja tendo calculado a resposta
internamente.

Grover() agora retorna "long result" em vez de "float elapsed" -- o
timing, que antes era medido dentro da propria funcao, migrou pra
grover.cpp (mesmo padrao que Shor()/shor.cpp ja usavam). grover.cpp
imprime "Found value: N" ou "Failed to find value (...)" comparando
contra search_value.

Verificado: 30/30 execucoes bem-sucedidas em testes manuais
(t_CPU/t_GPU/t_HYBRID, 10 e 12 qubits, Windows e WSL com GPU real) --
Grover e bem mais confiavel que Shor aqui (iteracao escolhida perto do
otimo). tests/smoke_test.sh continua so conferindo "sem crash" pra
grover.out, mesmo raciocinio ja aplicado a shor.out (continua
probabilistico por natureza, apesar da taxa alta observada). make test
(66/66 + smoke test) sem regressao nos dois ambientes.

+++ include/algorithms/lib_grover.h
@@ -14,7 +14,9 @@ using namespace std;
 
 // Busca "search_value" num espaço de qubits-1 qubits (o qubit 0 é
 // ancilla) via amplificação de amplitude — ver docs/05-algoritmo-grover.md.
-float Grover(long qubits, long search_value, int type, int thread_count, int cpu_region_bits, int cpu_coalesced_bits, int gpu_count, int gpu_region_bits, int gpu_coalesced_bits, int block_size, int repeat_count);
+// Devolve o valor medido do registrador de busca (com alta probabilidade
+// igual a search_value, mas não garantido — Grover é probabilístico).
+long Grover(long qubits, long search_value, int type, int thread_count, int cpu_region_bits, int cpu_coalesced_bits, int gpu_count, int gpu_region_bits, int gpu_coalesced_bits, int block_size, int repeat_count);
 
 // Z multi-controlado sobre todos os qubits de busca — parte do difusor.
 string ControledZ(int qubits);
